# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EarScope API",
        "description": "Dokumentasi API untuk EarScope",
        "version": "1.0"
    },
    "paths": {
...
        "/v1/videos": {
            "get": {
                "tags": [
                    "Videos"
                ],
                "summary": "Get all videos",
                "operationId": "2dd39694f233ae275064d4e27e73aea0",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter videos by created date (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all videos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "raw_video_url": {
                                                "type": "string"
                                            },
                                            "processed_video_url": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Upload a pair of videos (raw and processed)",
                "operationId": "766f06dc2935b23771613a91839bb1a1",
                "requestBody": {
                    "description": "Upload raw and processed videos",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "raw_video",
                                    "processed_video"
                                ],
                                "properties": {
                                    "raw_video": {
                                        "description": "Raw video file",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "processed_video": {
                                        "description": "Processed video file with bounding box",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "hasil_diagnosis": {
                                        "description": "Diagnosis result of the video (optional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Videos uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "raw_video_path": {
                                                    "type": "string"
                                                },
                                                "processed_video_path": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "hasil_diagnosis": {
                                                    "description": "Diagnosis result of the video",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/{videoId}/assign/{userId}": {
            "post": {
                "tags": [
                    "Videos"
                ],
                "summary": "Assign a video to a user",
                "operationId": "ea5bac4dbe008179d0e71352b804c1ef",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video assigned successfully"
                    },
                    "404": {
                        "description": "Video or User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/videos/{videoId}": {
            "patch": {
                "tags": [
                    "Videos"
                ],
                "summary": "Unassign a video",
                "operationId": "3ab86baf997ffc3d718acfe4a42cce06",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Video unassigned successfully"
                    },
                    "404": {
                        "description": "Video not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/videos/{videoId}/keterangan": {
            "patch": {
                "tags": [
                    "Videos"
                ],
                "summary": "Update keterangan video",
                "operationId": "22c90b7440129cb3b27fb35efe70d587",
                "parameters": [
                    {
                        "name": "videoId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "keterangan": {
                                        "type": "string",
                                        "example": "Diagnosis manual"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token di kolom ini dengan format Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}